May 27th
Finished day 8 (I think?). Didn't time myself because I read it, then did housework and what not, then ate dinner and coded it. It was definitely under half an hour, but it also sat in the back of my mind for awhile and I don't think I can state the accurate amount of time I coded it.