Finished day 12: May 31st
It probably took 45 min. 

I wrote the X=[1, 2] case super quickly (in maybe 5 minutes), and it had an error in it, but I didn't realize. However, the error hid itself in the non-general case. When I moved to a general case, that error couldn't hide itself anymore, so it took a long time to find it. 

The error was when I was recursively passing in my function, I was also taking some elements off of X. This was fixed and the reason I was doing this was better handled with some if/else statements and proper base cases.